@@ -4,42 +4,42 @@
   "private": true,
   "proxy": "https://localhost:5001",
   "dependencies": {
-    "@emotion/react": "^11.4.1",
-    "@emotion/styled": "^11.3.0",
-    "@mui/icons-material": "^5.0.3",
-    "@mui/material": "^5.0.3",
-    "axios": "^0.21.1",
+    "@emotion/react": "^11.10.5",
+    "@emotion/styled": "^11.10.5",
+    "@mui/icons-material": "^5.10.15",
+    "@mui/material": "^5.10.15",
+    "axios": "^1.2.0",
     "fontsource-roboto": "^3.1.5",
     "lodash": "^4.17.21",
-    "mobx": "^6.3.2",
-    "mobx-react": "^7.2.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "^4.0.3",
-    "uuid": "^8.3.2"
+    "mobx": "^6.7.0",
+    "mobx-react": "^7.6.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-router-dom": "^6.4.3",
+    "react-scripts": "^5.0.1",
+    "uuid": "^9.0.0"
   },
   "devDependencies": {
-    "@types/faker": "^5.5.6",
-    "@types/jest": "^26.0.23",
-    "@types/lodash": "^4.14.170",
-    "@types/node": "^15.12.3",
-    "@types/react": "^17.0.11",
-    "@types/react-dom": "^17.0.8",
-    "@types/react-router-dom": "^5.1.7",
-    "@types/uuid": "^8.3.0",
-    "ajv": "^8.6.0",
+    "@types/faker": "^6.6.8",
+    "@types/jest": "^29.2.3",
+    "@types/lodash": "^4.14.190",
+    "@types/node": "^18.11.9",
+    "@types/react": "^18.0.25",
+    "@types/react-dom": "^18.0.9",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/uuid": "^8.3.4",
+    "ajv": "^8.11.2",
     "cross-env": "^7.0.3",
-    "eslint": "^7.28.0",
-    "eslint-config-react-app": "^6.0.0",
-    "eslint-plugin-flowtype": "^5.7.2",
-    "eslint-plugin-import": "^2.23.4",
-    "eslint-plugin-jsx-a11y": "^6.4.1",
-    "eslint-plugin-react": "^7.24.0",
-    "faker": "^5.5.3",
-    "json-server": "^0.16.3",
-    "nan": "^2.14.2",
-    "typescript": "^4.3.2"
+    "eslint": "^8.28.0",
+    "eslint-config-react-app": "^7.0.1",
+    "eslint-plugin-flowtype": "^8.0.3",
+    "eslint-plugin-import": "^2.26.0",
+    "eslint-plugin-jsx-a11y": "^6.6.1",
+    "eslint-plugin-react": "^7.31.11",
+    "faker": "^6.6.6",
+    "json-server": "^0.17.1",
+    "nan": "^2.17.0",
+    "typescript": "^4.9.3"
   },
   "eslintConfig": {
     "extends": "react-app"
